+++ testing/test_pytest_cli.py
@@ -1,4 +1,5 @@
 from xdoctest.utils import util_misc
+import sys
 from xdoctest import utils
 
 
@@ -34,7 +35,7 @@ def test_simple_pytest_cli():
     temp_module = util_misc.TempModule(module_text)
     modpath = temp_module.modpath
 
-    info = cmd('pytest --xdoctest ' + modpath)
+    info = cmd(sys.executable + ' -m pytest --xdoctest ' + modpath)
     print(info['out'])
     assert info['ret'] == 0
 
@@ -44,11 +45,11 @@ def test_simple_pytest_import_error_cli():
     This test case triggers an excessively long callback in xdoctest <
     dev/0.15.7
 
-    xdoctest ~/code/xdoctest/testing/test_pytest_errors.py test_simple_pytest_import_error_cli
+    xdoctest ~/code/xdoctest/testing/test_pytest_cli.py test_simple_pytest_import_error_cli
 
     import sys, ubelt
     sys.path.append(ubelt.expandpath('~/code/xdoctest/testing'))
-    from test_pytest_errors import *  # NOQA
+    from test_pytest_cli import *  # NOQA
     """
     module_text = utils.codeblock(
         '''
@@ -66,7 +67,7 @@ def test_simple_pytest_import_error_cli():
             """
         ''')
     temp_module = util_misc.TempModule(module_text, modname='imperr_test_mod')
-    command = 'pytest -v -s --xdoctest-verbose=3 --xdoctest ' + temp_module.dpath
+    command = sys.executable + ' -m pytest -v -s --xdoctest-verbose=3 --xdoctest ' + temp_module.dpath
     print(command)
     info = cmd(command)
     print(info['out'])
@@ -93,10 +94,10 @@ def test_simple_pytest_syntax_error_cli():
             """
         ''')
     temp_module = util_misc.TempModule(module_text)
-    info = cmd('pytest --xdoctest ' + temp_module.dpath)
+    info = cmd(sys.executable + ' -m pytest --xdoctest ' + temp_module.dpath)
     print(info['out'])
 
-    info = cmd('pytest --xdoctest ' + temp_module.modpath)
+    info = cmd(sys.executable + ' -m pytest --xdoctest ' + temp_module.modpath)
     print(info['out'])
 
 
@@ -111,7 +112,7 @@ def test_simple_pytest_import_error_no_xdoctest():
             print('hello world')
         ''')
     temp_module = util_misc.TempModule(module_text)
-    info = cmd('pytest ' + temp_module.modpath)
+    info = cmd(sys.executable + ' -m pytest ' + temp_module.modpath)
     print(info['out'])
 
     info = cmd('pytest ' + temp_module.dpath)
@@ -130,9 +131,9 @@ def test_simple_pytest_syntax_error_no_xdoctest():
             print('hello world')
         ''')
     temp_module = util_misc.TempModule(module_text)
-    info = cmd('pytest ' + temp_module.modpath)
+    info = cmd(sys.executable + ' -m pytest ' + temp_module.modpath)
     print(info['out'])
 
-    info = cmd('pytest ' + temp_module.dpath)
+    info = cmd(sys.executable + ' -m pytest ' + temp_module.dpath)
     print(info['out'])
     # assert info['ret'] == 0